+++ src/main/java/com/twentyone/steachserver/domain/gpt/service/GPTService.java
@@ -1,4 +1,4 @@
-package com.twentyone.steachserver.domain.gpt;
+package com.twentyone.steachserver.domain.gpt.service;
 
 public interface GPTService {
     String getChatGPTResponse(String gptMessage);
